final 16 bit struct

+++ include/core/types.hpp
@@ -7,7 +7,7 @@ namespace raijin
     struct Order
     {
         uint64_t order_id;
-        uint64_t price_tick;
+        uint32_t price_tick;
         uint32_t volume;
     };
 }
